+++ Psado/Private/Remove-PSADOApi.ps1
@@ -22,7 +22,7 @@ Function Remove-PSADOApi {
     .NOTES
     Private function
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>