+++ Psado/Public/Get-PSADOBuild.ps1
@@ -38,7 +38,7 @@ function Get-PSADOBuild {
 
     .NOTES
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>