+++ Psado/Public/Get-PSADORelease.ps1
@@ -39,7 +39,7 @@ function Get-PSADORelease {
 
     .NOTES
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>