+++ Psado/Public/Invoke-PSADOBuild.ps1
@@ -30,7 +30,7 @@ function Invoke-PSADOBuild {
 
     .NOTES
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>